# api/openapi.yaml
openapi: 3.0.3
info:
  title: CodeForge API
  description: Remote AI code task runner. Submit code tasks, monitor progress via SSE, and create pull requests.
  version: 1.0.0
  contact:
    name: CodeForge
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /api/v1/webhooks/github:
    post:
      summary: GitHub webhook receiver for PR reviews
      operationId: githubWebhook
      security: []
      tags: [Webhooks]
      description: |
        Receives GitHub pull_request webhook events and automatically creates pr_review tasks.
        Verified via HMAC-SHA256 signature (X-Hub-Signature-256 header).
        Handles actions: opened, synchronize, reopened. Skips drafts unless configured.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: GitHub pull_request webhook payload
      responses:
        "201":
          description: Review task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: created
                  task_id:
                    type: string
                    format: uuid
        "200":
          description: Event ignored (not a handled action or draft)
        "401":
          description: Invalid webhook signature
  /api/v1/webhooks/gitlab:
    post:
      summary: GitLab webhook receiver for MR reviews
      operationId: gitlabWebhook
      security: []
      tags: [Webhooks]
      description: |
        Receives GitLab Merge Request Hook events and automatically creates pr_review tasks.
        Verified via X-Gitlab-Token header constant-time comparison.
        Handles actions: open, update, reopen. Skips draft/WIP unless configured.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: GitLab merge_request webhook payload
      responses:
        "201":
          description: Review task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: created
                  task_id:
                    type: string
                    format: uuid
        "200":
          description: Event ignored (not a handled action or draft/WIP)
        "401":
          description: Invalid webhook token
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...